상품 api 수정

+++ src/products/products.service.ts
@@ -5,12 +5,18 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Products } from './entities/products.entity';
 import { ProductsException } from './exceptions/products-exceptions';
+import { Request } from 'express';
+import { TokenPayload } from 'src/auth/interfaces/token-payload.interface';
+import { AuthService } from 'src/auth/services/auth.service';
+import { AuthException } from 'src/auth/exceptions/auth-exceptions';
 
 @Injectable()
 export class ProductService {
   constructor(
     @InjectRepository(Products)
     private productRepository: Repository<Products>,
+
+    private readonly authService: AuthService,
   ) {}
 
   private getProducts() {
@@ -44,6 +50,18 @@ export class ProductService {
     return product;
   }
 
+  private async verifyProductOwner(
+    id: number,
+    request: Request,
+  ): Promise<void> {
+    const payload = request['user'] as TokenPayload;
+    const findUser = await this.authService.findUserByEmail(payload.email);
+    const findProduct = await this.findProduct(id);
+    if (findUser.user_id !== findProduct.brand.user.user_id) {
+      throw new AuthException(AuthException.LOGIN_FAIL, HttpStatus.NOT_FOUND);
+    }
+  }
+
   async create(createProductDto: CreateProductDto) {
     return this.productRepository.save(createProductDto);
   }
@@ -58,13 +76,18 @@ export class ProductService {
     return product;
   }
 
-  async update(id: number, updateProductDto: UpdateProductDto) {
+  async update(
+    id: number,
+    updateProductDto: UpdateProductDto,
+    request: Request,
+  ) {
+    await this.verifyProductOwner(id, request);
     updateProductDto.updated_at = new Date();
     return await this.productRepository.update(id, updateProductDto);
   }
 
-  async remove(id: number) {
-    await this.findProduct(id);
+  async remove(id: number, request: Request) {
+    await this.verifyProductOwner(id, request);
     return await this.productRepository.delete(id);
   }
 }
